// src/types/provider.types.ts
export interface EditPolicyFormData {
  policyName: string;
  policyLink: string;
  contactName: string;
  contactPhone: string;
  contactEmail: string;
  insuranceCompany: string;
  effectiveDate: string;
  expirationDate: string;
  coverageLimit: string;
  premium: string;
  coverageBasis: string;
  policyHistory: string;
  retroactiveDate: string;
  bankruptcy: string;
  deductible: string;
  lossRun: string;
}